wip: refactor font creation
